Added tests, prototypes and refactored
Added a Plain type to the MatchData enum for matches which require no metadata. Also added the function in the scoring to generate bruteforce matches (which don't have metadata). Added constants used in the scoring process to the top of the file for calculating guesses in future.

Also, added tests for the factorial function, and fixed some issues in existing code that weren't exposed because the functions weren't called.