Added fancy regex and cleaned up some stuff.
Addd fancy regex cause it allows referencing previous captures for the repeat_match. Also, refactored the build script as the order elements appear in the map is important in the dropbox implementation.

Additionally removed unnecessary #include! from the lib file which removed a lot of warnings and some of the initial code for the repeat matcher and spatial matcher has been started.